refactor: unify getUserFromRequest
