+++ src/app/api/stories/bulk/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from "next/server";
 import { caslGuardWithPolicies } from "@/core/auth/casl.guard";
 import { prisma } from "@/core/prisma";
 import { StoryType, DifficultyLevel, ContentStatus } from "@prisma/client";
-import { getUserFromRequest } from "@/lib/auth";
+import { getUserFromRequest } from "@/core/auth/getUser";
 
 // PUT /api/stories/bulk - bulk update stories
 export async function PUT(request: NextRequest) {
